% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alhomodel.R
\name{alhomodel}
\alias{alhomodel}
\title{Conditional Maximum Likelihood (Alho 1990)}
\usage{
alhomodel(Y1, Y2, X)
}
\arguments{
\item{Y1}{a vector of capture histories.}

\item{Y2}{a vector of capture histories.}

\item{X}{a matrix of coefficients.}
}
\value{
List with components "N" and "beta". 
\describe{
\item{"N"}{is a single estimate of the population size.}
\item{"beta"}{is a matrix of beta estimates.}
}
}
\description{
This function performs the algorithm in Alho 1990 and returns an estimate for
beta and N.
}
\examples{
NA

}
