% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiParametricBRCRC.R
\name{BLRCRsolver2}
\alias{BLRCRsolver2}
\title{A placeholder function using roxygen}
\usage{
BLRCRsolver2(
  y,
  x,
  priorb,
  priorB,
  priornu0,
  priorLAMBDA0,
  priorkappa0,
  priorMU0,
  aalpha,
  balpha,
  Kstar,
  rhomethod = 2,
  samples = 100
)
}
\arguments{
\item{txt}{An optional character variable, defaults to \sQuote{world}}
}
\value{
Nothing is returned but as a side effect output is printed
}
\description{
This function shows a standard text on the console. In a time-honoured
tradition, it defaults to displaying \emph{hello, world}.
}
\examples{
hello2()
hello2("and goodbye")
}
