% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BLRCRsolvermissing}
\alias{BLRCRsolvermissing}
\title{Bayesian Logistic Regression Capture-Recapture with Missing Covariate Values}
\usage{
BLRCRsolvermissing(
  formula,
  df,
  covmethod = "bootstrap",
  covsupport = "unique",
  coefprior = "normal",
  bprior = NULL,
  Bprior = 1,
  LBprior = 1,
  alphabayesbootstrap = 0,
  alphaomega = NULL,
  aomega = 0.25,
  bomega = 0.25,
  Homega = 1L,
  samples = 1000L,
  burnin = 0L,
  thinning = 1L
)
}
\arguments{
\item{formula}{an object of class "formula".}

\item{df}{data containing capture histories and individual covariates.}

\item{covmethod}{specify a method for the distribution of the covariates.
Options include "bootstrap" and "empirical".}

\item{covsupport}{specify a method for generating the support of the bootstrap distribution.
Options include "all" and "unique".}

\item{coefprior}{specify a prior for the logistic regression coefficients.
Options include "normal", "normalhier", "normalhiervaronly", and "lasso".}

\item{bprior}{vector of hyperparameters for the mean of the coefficient prior. If a single 
value is used, it will apply this value for all coefficients. If "normalhier" is selected, it will use
these values as the coefficient prior on b.}

\item{Bprior}{value to assign the diagonals of the covariance matrix B for the observed variables.}

\item{LBprior}{value to assign the diagonals of the covariance matrix B for the latent variables.}

\item{alphaomega}{a value for the concentration parameter. Leave blank to apply a prior distribution
to the concentration parameter of Gamma(aomega,bomega).}

\item{aomega}{a value for the hyperparameter on the concentration parameter.}

\item{bomega}{a value for the hyperparameter on the concentration parameter.}

\item{Homega}{the number of latent intercepts}

\item{samples}{the number of MCMC samples to draw and return from the posterior.}

\item{burnin}{the number of MCMC samples to "burn-in."}

\item{thinning}{the number of MCMC samples to "thin".}
}
\value{
\describe{
\item{"samples"}{is the number of samples.} 
\item{"H"}{is the number of covariate values after transformations} 
\item{"n"}{is the number of observed individuals in the data.}
\item{"N"}{is a sample of population sizes from the posterior distribution.}
\item{"Xmis"}{is a sample of missing covariate values. One of each covariate is sampled and saved
at each sampling occasion.}
\item{"alphaomega"}{is a sample of alphaomega values from the posterior.}
\item{"Beta"}{is a sample of JxH+Homega matrices of beta coefficients.}
\item{"Betacolnames"}{The names of the covariates for each column of the beta coefficients.}
\item{"PSIMATRIX"}{}
\item{"PSICOUNTS"}{}
\item{"PHIomega"}{}
\item{"omegalabels"}{}
\item{"priorBdiag"}{}
\item{"probcaptured"}{}
\item{"loglikelihood"}{}
\item{"BIC"}{}
}
}
\description{
This function computes MCMC samples of N using the the Bayesian Logistic
Regression Capture-Recapture model but can also handle missing covariate values.
This method only allows for "bootstrap" and "empirical" covariate methods.
}
\examples{
#Create the Data
mybeta = matrix(c(-2,-1,1,
                  -2,1,-1,
                  -2,-1,1,
                  -2,1,-1),nrow=4,byrow=TRUE)
                  
mynormprobs=c(0.3,0.4,0.3)

mynormmeans=matrix(c(2,2,
                     0,0,
                     -2,-2),ncol=2,byrow=TRUE)
                     
mynormsigma=matrix(c(.5,.45,.45,.5,
                      1,0,0,1,
                      .5,-.35,-.35,.5),nrow=3,byrow=TRUE)

myCRdata = multdatasimulator(2000,mybeta,mynormprobs,mynormmeans,mynormsigma,
                           missing=TRUE)
                           
#Run the Algorithm
myformula = cbind(y1,y2,y3,y4)~x1+x2

CRresults=BLRCRsolvermissing(myformula,df=myCRdata,Homega=1,
                             Bprior=1,LBprior=1,
                             covsupport="unique",
                             samples = 1000,burnin=10,thinning = 10)


}
