% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bootstrapRCR}
\alias{bootstrapRCR}
\title{Semi-Parametric Bayesian Logistic Regression Capture-Recapture}
\usage{
bootstrapRCR(y, x, beta)
}
\arguments{
\item{y}{a nxJ matrix of capture patterns.}

\item{x}{a nxH matrix of covariates.}

\item{beta}{a (H+1) vector of prior means for the coefficients}
}
\value{
List with components "N" and "beta". 
\describe{
\item{"N"}{is a single estimate of the population size.}
\item{"beta"}{is a matrix of beta estimates.}
}
}
\description{
This function computes MCMC samples of N using the the Bayesian Logistic
Regression Capture-Recapture model.
}
\examples{
#Generate the Data
mybeta = matrix(c(-2,-1,1,
                  -2,1,-1,
                  -2,-1,1,
                  -2,1,-1),nrow=4,byrow=TRUE)
                  

}
